2.0.0.beta06: all temporal tests appears correctly adaptive
- TrialBasic:
stim_value is calculated at runtime starting from magnitude and the extra_param properties (e.g. in bis indicates whether the 2nd stim is before or after 500ms)

open var correct_answer:Int is now not passed as a constructor parameter. happens only in TrialMMD, TrialTFI
in all other subclasses is calculated in the updateTrial method.

calculated adaptive value is coerced to be between 0 and range

added onItemSelectedListener for spTrialManager in order to update allowed conditions according to trial management

some optimization in strings.xml

AdaptiveTrialsManager now coerces magnitude between 0 and range

spManager in SubjectBasicDialogFragment has now onItelSelectedListener to modify available condition w/wo apaptive trials

+++ psysuitecore/src/main/java/iit/uvip/psysuite/core/tests/temporalbinding/BindingsConstants.kt
@@ -4,19 +4,19 @@ class BindingsConstants {
 
     companion object {
 
-        @JvmStatic val balshSD:List<Pair<Long, String>> = listOf(
-            Pair(200L, "200"),
-            Pair(250L, "250"),
-            Pair(300L, "300"))
+        @JvmStatic val balshSD:List<Pair<Float, String>> = listOf(
+            Pair(200.0F,    "200"),
+            Pair(250.0F,    "250"),
+            Pair(300.0F,    "300"))
 
-        @JvmStatic val unbalSD:List<Pair<Long, String>> = listOf(
-            Pair(50L,  "50"),
-            Pair(100L, "100"),
-            Pair(200L, "200"),
-            Pair(300L, "300"),
-            Pair(400L, "400"),
-            Pair(800L, "800"),
-            Pair(1200L,"1200"))
+        @JvmStatic val unbalSD:List<Pair<Float, String>> = listOf(
+            Pair(50.0F,     "50"),
+            Pair(100.0F,    "100"),
+            Pair(200.0F,    "200"),
+            Pair(300.0F,    "300"),
+            Pair(400.0F,    "400"),
+            Pair(800.0F,    "800"),
+            Pair(1200.0F,   "1200"))
 
         @JvmStatic val WN_FIRSTSTIM_INTERVAL   = 1000L
         @JvmStatic val STIM_DURATION_INF       = 1000L
@@ -25,14 +25,12 @@ class BindingsConstants {
         @JvmStatic val ISI                     = 1000L
         @JvmStatic val ISI_INF                 = 2000L // distance between stimuli onsets
 
-
         // unimodal
         @JvmStatic val TYPE_A      = 1
         @JvmStatic val TYPE_T      = 2
         @JvmStatic val TYPE_V      = 3
 
         // bimodal
-
         @JvmStatic val TYPE_AT     = 12
         @JvmStatic val TYPE_AV     = 13
         @JvmStatic val TYPE_TV     = 23
